Move wait on instance method to base provider.
