+++ ipa/ipa_provider.py
@@ -25,6 +25,7 @@ import json
 import logging
 import os
 import shlex
+import time
 
 import pytest
 
@@ -371,6 +372,13 @@ class IpaProvider(object):
             desc=self.desc,
             test_log=self.log_file
         )
+
+    def _wait_on_instance(self, state):
+        """Wait until instance is in given state."""
+        current_state = 'Undefined'
+        while state != current_state:
+            time.sleep(10)
+            current_state = self._get_instance_state()
 
     def hard_reboot_instance(self):
         """Stop then start the instance."""
